backend/requirements.txt
# === FIXED requirements.txt ===
# spaCy upgraded to 3.8.x which supports typer>=0.12 (compatible with fastapi-cli)

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0

# --- Database ---
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# --- NLP / AI ---
spacy==3.8.3
transformers==4.41.1
torch==2.3.0
scikit-learn==1.5.0

# --- PDF Processing ---
pymupdf==1.24.3

# --- Data Collection ---
beautifulsoup4==4.12.3
requests==2.32.2

# --- Data Handling ---
pandas==2.2.2
python-dotenv==1.0.1

# --- Utilities ---
pydantic==2.7.1
python-multipart==0.0.9